refactor: migrate workflow related methods in workflows store
