+++ packages/frontend/editor-ui/src/app/components/ActivationModal.vue
@@ -18,11 +18,18 @@ import {
 } from '../constants';
 
 import { N8nButton, N8nCheckbox, N8nText } from '@n8n/design-system';
+import {
+	createWorkflowDocumentId,
+	useWorkflowDocumentStore,
+} from '../stores/workflowDocument.store';
 
 const checked = ref(false);
 
 const executionsStore = useExecutionsStore();
 const workflowsStore = useWorkflowsStore();
+const workflowDocumentStore = computed(() =>
+	useWorkflowDocumentStore(createWorkflowDocumentId(workflowsStore.workflowId)),
+);
 const nodeTypesStore = useNodeTypesStore();
 const uiStore = useUIStore();
 const router = useRouter();
@@ -31,7 +38,9 @@ const i18n = useI18n();
 const modalTitle = computed(() => i18n.baseText('activationModal.workflowPublished'));
 
 const triggerContent = computed(() => {
-	const foundTriggers = getActivatableTriggerNodes(workflowsStore.workflowTriggerNodes);
+	const foundTriggers = getActivatableTriggerNodes(
+		workflowDocumentStore.value.workflowTriggerNodes,
+	);
 	if (!foundTriggers.length) {
 		return '';
 	}
