+++ packages/frontend/editor-ui/src/app/composables/useNodeHelpers.test.ts
@@ -33,6 +33,7 @@ vi.mock('@/app/stores/workflowDocument.store', async () => {
 			pinData: {},
 			usedCredentials: mockDocumentStoreUsedCredentials,
 			allNodes: [],
+			workflowTriggerNodes: [],
 			getNodeByName: vi.fn(),
 			setNodeIssue: vi.fn(),
 			updateNodeProperties: vi.fn(),
@@ -81,6 +82,7 @@ describe('useNodeHelpers()', () => {
 				settings: {},
 				usedCredentials: mockDocumentStoreUsedCredentials,
 				allNodes: [],
+				workflowTriggerNodes: [],
 				getNodeByName: vi.fn().mockReturnValue(node),
 				setNodeIssue: vi.fn(),
 				updateNodeProperties: vi.fn(),
