+++ packages/frontend/editor-ui/src/app/composables/usePinnedData.test.ts
@@ -30,6 +30,7 @@ const getNodeType = vi.fn();
 vi.mock('@/app/stores/nodeTypes.store', () => ({
 	useNodeTypesStore: vi.fn(() => ({
 		getNodeType,
+		getAllNodeTypes: vi.fn().mockReturnValue([]),
 	})),
 }));
