+++ packages/frontend/editor-ui/src/app/composables/usePushConnection/handlers/workflowSettingsUpdated.test.ts
@@ -14,6 +14,7 @@ const { mockWorkflowDocumentStore } = vi.hoisted(() => ({
 		allNodes: [],
 		name: '',
 		settings: {},
+		workflowTriggerNodes: [],
 		mergeSettings: vi.fn(),
 		setChecksum: vi.fn(),
 		getPinDataSnapshot: vi.fn().mockReturnValue({}),
