+++ packages/frontend/editor-ui/src/experiments/resourceCenter/components/__tests__/ResourceFeatureCard.test.ts
@@ -6,7 +6,14 @@ import ResourceFeatureCard from '../ResourceFeatureCard.vue';
 import type { ResourceItem } from '../../data/resourceCenterData';
 
 vi.mock('@/app/stores/nodeTypes.store', () => ({
-	useNodeTypesStore: () => ({ getNodeType: vi.fn(() => null) }),
+	useNodeTypesStore: () => ({
+		getNodeType: vi.fn(() => null),
+		getAllNodeTypes: vi.fn().mockReturnValue({
+			nodeTypes: {},
+			init: async () => {},
+			getByNameAndVersion: () => undefined,
+		}),
+	}),
 }));
 
 const readyToRunItem: ResourceItem = {
