+++ packages/frontend/editor-ui/src/features/ai/assistant/composables/useNodeMention.test.ts
@@ -24,6 +24,7 @@ vi.mock('@/features/ndv/shared/ndv.store', () => ({
 const { mockWorkflowDocumentStore } = vi.hoisted(() => ({
 	mockWorkflowDocumentStore: {
 		allNodes: [] as INodeUi[],
+		workflowTriggerNodes: [] as INodeUi[],
 		name: '',
 		settings: {},
 		getPinDataSnapshot: () => ({}),
