+++ packages/frontend/editor-ui/src/features/execution/executions/components/workflow/WorkflowExecutionsLandingPage.vue
@@ -9,6 +9,10 @@ import WorkflowExecutionsInfoAccordion from './WorkflowExecutionsInfoAccordion.v
 import { useI18n } from '@n8n/i18n';
 
 import { N8nButton, N8nHeading, N8nText } from '@n8n/design-system';
+import {
+	createWorkflowDocumentId,
+	useWorkflowDocumentStore,
+} from '@/app/stores/workflowDocument.store';
 const router = useRouter();
 const route = useRoute();
 const locale = useI18n();
@@ -16,9 +20,12 @@ const locale = useI18n();
 const workflowId = useInjectWorkflowId();
 const uiStore = useUIStore();
 const workflowsStore = useWorkflowsStore();
+const workflowDocumentStore = computed(() =>
+	useWorkflowDocumentStore(createWorkflowDocumentId(workflowsStore.workflowId)),
+);
 
 const executionCount = computed(() => workflowsStore.currentWorkflowExecutions.length);
-const containsTrigger = computed(() => workflowsStore.workflowTriggerNodes.length > 0);
+const containsTrigger = computed(() => workflowDocumentStore.value.workflowTriggerNodes.length > 0);
 
 function onSetupFirstStep(): void {
 	const resolvedWorkflowId = workflowId.value || route.params.workflowId;
