+++ packages/frontend/editor-ui/src/features/ndv/panel/components/NDVSubConnections.test.ts
@@ -55,9 +55,10 @@ const getNodeType = vi.fn();
 let mockGetNodeByName: Mock<(name: string) => INodeUi | null> = vi.fn(() => node);
 
 vi.mock('@/app/stores/nodeTypes.store', () => ({
-	useNodeTypesStore: vi.fn(() => ({
+	useNodeTypesStore: () => ({
 		getNodeType,
-	})),
+		getAllNodeTypes: () => ({}),
+	}),
 }));
 
 vi.mock('@/app/stores/workflowDocument.store', async (importActual) => ({
