+++ packages/frontend/editor-ui/src/features/ndv/parameters/components/ParameterInputList.test.ts
@@ -19,6 +19,8 @@ vi.mock('@/app/stores/workflowDocument.store', async (importOriginal) => ({
 		name: '',
 		settings: {},
 		getPinDataSnapshot: () => ({}),
+		workflowTriggerNodes: [],
+		allNodes: [],
 	}),
 }));
 
@@ -116,6 +118,8 @@ const workflowDocumentStoreMock = {
 	name: '',
 	settings: {},
 	getPinDataSnapshot: vi.fn().mockReturnValue({}),
+	workflowTriggerNodes: [],
+	allNodes: [],
 };
 
 const renderComponent = createComponentRenderer(ParameterInputList, {
