+++ packages/frontend/editor-ui/src/features/shared/editors/plugins/codemirror/completions/utils.test.ts
@@ -31,6 +31,7 @@ const { mockWorkflowDocumentStore } = vi.hoisted(() => ({
 		name: '',
 		settings: {},
 		getPinDataSnapshot: () => ({}),
+		workflowTriggerNodes: [],
 	},
 }));
