+++ packages/frontend/editor-ui/src/features/shared/nodeCreator/nodeCreator.store.ts
@@ -143,7 +143,7 @@ export const useNodeCreatorStore = defineStore(STORES.NODE_CREATOR, () => {
 	}: ToggleNodeCreatorOptions) {
 		if (!nodeCreatorView) {
 			nodeCreatorView =
-				workflowsStore.workflowTriggerNodes.length > 0
+				workflowDocumentStore.value.workflowTriggerNodes.length > 0
 					? REGULAR_NODE_CREATOR_VIEW
 					: TRIGGER_NODE_CREATOR_VIEW;
 		}
