+++ packages/frontend/editor-ui/src/features/shared/nodeCreator/views/viewsData.test.ts
@@ -45,6 +45,11 @@ vi.mock('@/app/stores/nodeTypes.store', () => ({
 	useNodeTypesStore: vi.fn(() => ({
 		getNodeType,
 		allLatestNodeTypes: [aiTransformNode, ...otherNodes],
+		getAllNodeTypes: vi.fn().mockReturnValue({
+			nodeTypes: {},
+			init: async () => {},
+			getByNameAndVersion: () => undefined,
+		}),
 	})),
 }));
