+++ packages/frontend/editor-ui/src/features/workflows/canvas/components/elements/nodes/CanvasNode.test.ts
@@ -16,6 +16,11 @@ vi.mock('@/app/stores/nodeTypes.store', () => ({
 			name: 'test',
 			description: 'Test Node Description',
 		})),
+		getAllNodeTypes: vi.fn().mockReturnValue({
+			nodeTypes: {},
+			init: async () => {},
+			getByNameAndVersion: () => undefined,
+		}),
 	})),
 }));
