+++ packages/frontend/editor-ui/src/features/workflows/workflowDiff/WorkflowDiffView.test.ts
@@ -98,6 +98,11 @@ vi.mock('@/app/stores/nodeTypes.store', () => ({
 	useNodeTypesStore: () => ({
 		loadNodeTypesIfNotLoaded: vi.fn().mockResolvedValue(undefined),
 		getNodeType: vi.fn().mockReturnValue({ name: 'Test', version: 1 }),
+		getAllNodeTypes: vi.fn().mockReturnValue({
+			nodeTypes: {},
+			init: async () => {},
+			getByNameAndVersion: () => undefined,
+		}),
 	}),
 }));
